refactor: move provider lifecycle helpers into run_internal
